F01: Add implementation of ProjetPartenaireRegionService
