+++ src/main/java/com/carto/sn/service/ProjetPartenaireRegionService.java
@@ -1,5 +1,12 @@
 package com.carto.sn.service;
 
-public interface ProjetPartenaireRegionService {
+import java.util.List;
 
+import com.carto.sn.entities.ProjetPartenaireRegion;
+
+public interface ProjetPartenaireRegionService {
+	public List <ProjetPartenaireRegion> findByIdProjet(Long idProjet);
+	public List <ProjetPartenaireRegion> findByIdPartenaire(Long idPartenaire);
+	public List <ProjetPartenaireRegion> findByIdRegion(Long idRegion);
+	public List <ProjetPartenaireRegion> tousLesProjetsPartenairesRegions();
 }
